+++ src/commands/from-selection.command.ts
@@ -7,6 +7,7 @@ import { Uri, window } from 'vscode';
 import { generateClass, runBuildRunner, runDartFormat } from '../index';
 import { getSelectedText, handleError, validateJSON } from '../lib';
 import { promptForBaseClassName, promptForTargetDirectory } from '../shared/user-prompts';
+import { hasObject } from '../utils';
 
 export const transformFromSelection = async (uri: Uri) => {
     const className = await promptForBaseClassName();
@@ -34,13 +35,14 @@ export const transformFromSelection = async (uri: Uri) => {
     }
 
     const json: string = await getSelectedText().then(validateJSON).catch(handleError);
-
+    const hasObj = hasObject(JSON.parse(json));
+    const model = new ClassNameModel(className);
     const config: Settings = {
-        model: new ClassNameModel(className),
+        model: model,
         targetDirectory: <string>targetDirectory,
         json: json,
         input: input,
-        targetDirectoryType: TargetDirectoryType.Standard,
+        targetDirectoryType: hasObj ? TargetDirectoryType.Compressed : TargetDirectoryType.Expanded,
     };
 
     if (!input.primaryConfiguration) {
@@ -52,7 +54,10 @@ export const transformFromSelection = async (uri: Uri) => {
     const settings = new Settings(config);
 
     await generateClass(settings).then((_) => {
-        runDartFormat(<string>targetDirectory, 'models');
+        const formattingTarget = hasObj ? model.directoryName : '';
+
+        runDartFormat(<string>targetDirectory, formattingTarget);
+
         if (input.generate && input.runBuilder) {
             runBuildRunner();
         }