+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
    - Added support for safe `JSON`. The generator can read and parse `json` and `jsonc` from any method.
    - Added support for multiple files conversion. The generator can read `**.json` and `**.jsonc` files from the tracked file or directory.
+   - Added a better method for writing files from the context or file explorer.
    - Improved conversion performance. 5X faster speed.
 
 ## 3.4.1